aggregator - decomposition - #2
